% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_stat.R
\name{ba_stat}
\alias{ba_stat}
\title{Bland-Altman statistics}
\usage{
ba_stat(
  data = stop("data must be specified"),
  var1 = stop("variable must be specified"),
  var2 = stop("variable must be specified"),
  group = NULL,
  alpha = 0.05
)
}
\arguments{
\item{data}{A data frame}

\item{var1}{1st variable to compare (unquoted)}

\item{var2}{2nd variable to compare (unquoted)}

\item{group}{grouping variable (unquoted)}

\item{alpha}{alpha level for the intervals}
}
\value{
A \link{tibble} with three variables n (number of observations), parameter and value is returned.
}
\description{
The function is computing all relevant Bland-Altman statistics, including bias, lower and upper
limits of agreement and their confidence limits.
}
\examples{
library(tidyr)
tbl <- temperature \%>\% pivot_wider(names_from = method, values_from = temperature)

# simple example
ba_stat(data = tbl, var1 = infrared, var2 = rectal)

## example with grouping
ba_stat(data = tbl, var1 = infrared, var2 = rectal, group = treatment) \%>\%
  pivot_wider(names_from = parameter, values_from = value)
}
\seealso{
\link{ba_plot}
}
