% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAplot-package.R
\docType{package}
\name{BAplot-package}
\alias{BAplot}
\alias{BAplot-package}
\title{BAplot: Compute and visualize Bland Altman statistics}
\description{
The package provides R functions to compute Bland-Altman statistics and to visualize those statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KonstantinLang/BAplot}
}

}
\author{
\strong{Maintainer}: Konstantin Lang \email{konstantin.lang1@bayer.com}

}
\keyword{internal}
