% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_temp_data.R
\name{gen_temp_data}
\alias{gen_temp_data}
\title{Generate rodents temperature data}
\usage{
gen_temp_data(seed = NULL)
}
\arguments{
\item{seed}{a single value, interpreted as an integer, or \code{NULL}}
}
\value{
\code{\link{tibble}} with columns
\item{animalID}{unique animal ID: 1-90, \code{integer}}
\item{treatment}{one of: healthy, untreated, low dose, mid dose, high dose, positive control, \code{factor}}
\item{method}{measurement method: rectal, infrared, \code{factor}}
\item{visit}{one of: baseline, visit1, visit2, visit3, end of treatment, \code{factor}}
\item{tempareture}{body temparature in °C, \code{numeric}}
}
\description{
Generate rodents temperature data
}
\examples{
temperature <- gen_temp_data()
}
