% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBA-package.R
\docType{package}
\name{ggBA-package}
\alias{ggBA}
\alias{ggBA-package}
\title{ggBA: Compute and Visualize Bland Altman Statistics}
\description{
R functions to compute Bland-Altman statistics and to visualize those statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KonstantinLang/ggBA}
}

}
\author{
\strong{Maintainer}: Konstantin Lang \email{konstantin.lang1@bayer.com}

}
\keyword{internal}
