% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature}
\alias{temperature}
\title{Rodent Temperature Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with
900 rows and 5 columns:
\describe{
\item{animalID}{unique animal ID: 1-90, \code{integer}}
\item{treatment}{one of: healthy, untreated, low dose, mid dose, high dose, positive control, \code{factor}}
\item{method}{measurement method: rectal, infrared, \code{factor}}
\item{visit}{one of: baseline, visit1, visit2, visit3, end of treatment, \code{factor}}
\item{tempareture}{body temparature in °C, \code{numeric}}
}
}
\usage{
temperature
}
\description{
A data set with rodent data comparing two different temperature measurement methods of healthy
and treated animals.
}
\keyword{datasets}
