% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_plot.R
\name{ba_plot}
\alias{ba_plot}
\title{Bland-Altman plot}
\usage{
ba_plot(
  data = stop("data must be specified"),
  var1 = stop("variable must be specified"),
  var2 = stop("variable must be specified"),
  label = NULL,
  group = NULL,
  colour = NULL,
  shape = NULL,
  xlab = "Average",
  ylab = "Difference",
  title = NULL,
  caption = NULL,
  alpha = 0.05
)
}
\arguments{
\item{data}{A data frame}

\item{var1}{1st variable to compare (unquoted)}

\item{var2}{2nd variable to compare (unquoted)}

\item{label}{data label (unquote)}

\item{group}{grouping variable used for faceting (unquoted)}

\item{colour}{colour asthetic for scatter points (unquoted)}

\item{shape}{shape asthetic for scatter points (unquoted)}

\item{xlab}{The text for the x-axis label}

\item{ylab}{The text for the y-axis label}

\item{title}{plot title}

\item{caption}{plot caption}

\item{alpha}{alpha level for the intervals}
}
\value{
\link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\description{
Plot Bland-Altman statistics
}
\examples{
library(tidyr)
tbl <- temperature \%>\% pivot_wider(names_from = method, values_from = temperature)

# simple example
ba_plot(data = tbl, var1 = infrared, var2 = rectal)

# with colors
ba_plot(data = tbl, var1 = infrared, var2 = rectal, colour = visit)

# with colors and faceting
ba_plot(data = tbl, var1 = infrared, var2 = rectal, group = treatment, colour = visit)
}
\seealso{
\link{ba_stat}
}
