% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_stat.R
\name{ba_stat}
\alias{ba_stat}
\title{Bland-Altman statistics}
\usage{
ba_stat(data, var1, var2, group = NULL, alpha = 0.05)
}
\arguments{
\item{data}{A data frame}

\item{var1}{1st variable to compare (unquoted)}

\item{var2}{2nd variable to compare (unquoted)}

\item{group}{grouping variable (unquoted)}

\item{alpha}{alpha level for the intervals}
}
\value{
A data frame with three variables n (number of observations), parameter and value is returned.
}
\description{
The function is computing all relevant Bland-Altman statistics, including bias, lower and upper
limits of agreement and their confidence limits.
}
\examples{
## simple example without grouping
ba_stat(data = iris, var1 = Petal.Length, var2 = Petal.Width)

## example with grouping
ba_stat(data = iris, var1 = Petal.Length, var2 = Petal.Width, group = Species)
}
\seealso{
\link{ba_plot}
}
